requirements.txt
fastapi
geoalchemy2
python-dotenv
shapely
sqlalchemy
alembic
asyncpg
psycopg2-binary